refactor: SectionService 삭제
