+++ src/main/java/nextstep/subway/domain/Line.java
@@ -61,7 +61,7 @@ public class Line extends BaseEntity {
 
     public void addSection(Section section) {
         sections.addSection(section);
-        section.addLine(this);
+        section.updateLine(this);
     }
 
     public void removeSection(Station station) {
